updates for previous PR
