+++ src/components/CategoryWiseServices/CategoryWiseServices.css
@@ -1,0 +1,10 @@
+.containerCard{
+    width: 80%;
+    margin: 10px;
+}
+@media(max-width: 500px) {
+    .containerCard{
+        width: 90%;
+        margin: 10px;
+    }
+}
