+++ src/components/Sidebar/Sidebar.js
@@ -3,21 +3,39 @@ import { connect } from "react-redux";
 import { Link } from 'react-router-dom';
 import 'antd/dist/antd.css';
 import './Sidebar.css'
-import { Layout, Menu, Icon, Spin } from "antd";
+import { Layout, Menu, Icon, Spin, Alert } from "antd";
 
-import { getAllCategories } from "../../redux/actions/categoryActions";
+import { getAllCategories } from "../../redux/actions/categoryServiceActions";
 
 const { Sider } = Layout;
 
 class SideBar extends Component {
 
+  state={
+    msg: null
+  }
+
   componentDidMount() {
     this.props.getAllCategories();
   }
-
+  componentDidUpdate(prevProps){
+    const {error} = this.props;
+    if(error !== prevProps.error){
+        if(error.id === "CATEGORIES_FETCH_FAIL"){
+            this.setState({
+                msg : error.msg
+            });
+        }
+        else{
+            this.setState({
+                msg : null
+            });
+        }			
+    }
+  } 
   render() {
-
-      const categoryList=!this.props.category.pending ? ( this.props.category.categories.map((category)=>{
+    const {msg}=this.state;
+    const categoryList=!this.props.categoryService.pending ? ( this.props.categoryService.categories.map((category)=>{
         return(
           <Menu.Item key={category._id}>
             <Icon type="form" />
@@ -33,20 +51,21 @@ class SideBar extends Component {
 
     return (
       <div className="wrapper">
-      <div style={{position: 'relative'}}>
-        <Sider
-          style={{ overflow: 'auto', height: '100vh', left: 0 }}
-          trigger={null}
-          collapsible
-          collapsedWidth={0}
-          collapsed={this.props.collapseProp}
-        >
-          <div className="logo" />
-          <Menu theme="dark" mode="inline">
-            { categoryList }
-          </Menu>
-        </Sider>
-      </div>
+        <div style={{position: 'relative'}}>
+          {msg ? <Alert message={msg} type='error' /> : null}
+          <Sider
+            style={{ overflow: 'auto', height: '100vh', left: 0 }}
+            trigger={null}
+            collapsible
+            collapsedWidth={0}
+            collapsed={this.props.collapseProp}
+          >
+            <div className="logo" />
+            <Menu theme="dark" mode="inline">
+              { categoryList }
+            </Menu>
+          </Sider>
+        </div>
       </div>
     );
   }
@@ -54,7 +73,7 @@ class SideBar extends Component {
 
 const mapStateToProps = state => ({
   isAuthenticated: state.auth.isAuthenticated,
-  category: state.category
+  categoryService: state.categoryService
 });
 
 export default connect(
