+++ src/redux/actions/vendorActions.js
@@ -3,7 +3,7 @@ import axios from "axios";
 import {returnErrors} from './errorActions';
 import { tokenConfig } from './authActions';
 
-const url = "https://secure-falls-92714.herokuapp.com";
+import {url} from '../../helper/url';
 
 export const registerBusiness = ({businessname,description,residence,phone}) => (dispatch, getState) => {
   const detail = {
