+++ src/redux/reducers/index.js
@@ -3,8 +3,7 @@ import vendorReducer from './vendorReducer';
 import AuthReducer from './authReducer';
 import ErrorReducer from './errorReducer';
 import UserReducer from './userReducer';
-import CategoryReducer from './categoryReducer';
-import ServiceReducer from './serviceReducer';
+import CategoryServiceReducer from './categoryServiceReducer';
 
 
 export default combineReducers({
@@ -12,6 +11,5 @@ export default combineReducers({
   error: ErrorReducer,
   user: UserReducer,
   vendor: vendorReducer,
-  category: CategoryReducer,
-  service: ServiceReducer
+  categoryService: CategoryServiceReducer
 });
